remove path filter for backend deploys
This makes it so that the backend just gets built/deployed
on every push to main/preview. Relying on path alone is
unreliable -- we should probably check based on branch +
path instead. Deploys are usually pretty fast on main
anyways, since the docker build layers should be cached
remotely if we deploy to preview first.